refactor: routes.py
